ref(Assets): rename `AssetCache` to `VmInstanceManager`
