+++ Assets/Gothic-UnZENity-Core/Scripts/Manager/VobHelper.cs
@@ -6,6 +6,7 @@ using GUZ.Core.Data;
 using GUZ.Core.Globals;
 using GUZ.Core.Properties;
 using GUZ.Core.Extensions;
+using GUZ.Core.Vm;
 using GVR.Core;
 using JetBrains.Annotations;
 using UnityEngine;
@@ -68,7 +69,7 @@ namespace GUZ.Core.Manager
             }
             else
             {
-                var sfxData = AssetCache.TryGetSfxData(soundName);
+                var sfxData = VmInstanceManager.TryGetSfxData(soundName);
 
                 if (sfxData == null)
                     return null;
